requirements.txt
python-dotenv==1.1.0
pandas==2.2.3
pydantic==2.11.3
json_repair==0.41.1
openai==1.74.0
joblib==1.4.2
toml==0.10.2
pyarrow==19.0.1
ipykernel==6.29.5
sentence-transformers==4.1.0
camelsplit==1.0.1
bm25s==0.2.12
faiss-cpu==1.10.0
openpyxl==3.1.5